% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp2rd.R
\name{sp2rd}
\alias{sp2rd}
\title{Animation of location to reduced dimension transitions}
\usage{
sp2rd(
  object,
  coords = c("Pos_X", "Pos_Y"),
  color_by = NULL,
  img_id = "sample_id",
  dimred = "UMAP",
  margin = 100
)
}
\arguments{
\item{object}{a \code{SingleCellExperiment} or \code{SpatialExperiment} object.}

\item{coords}{the name of the x and y coordinates}

\item{color_by}{how to color the points}

\item{img_id}{the name of the image ID entry}

\item{dimred}{which dimensional reduction to use}

\item{margin}{padding between individual images}
}
\value{
a ggplot object
}
\description{
TODO
}
\examples{
# TODO

}
\author{
Nils Eling (\email{nils.eling@dqbm.uzh.ch})
}
